Implement routing to c23s-ui/master-ui login
#14592 - Single UI 1 : As a developer I will create a single UI interface to login

+++ client/src/app/security/shared/authentication.service.ts
@@ -43,10 +43,17 @@ export class AuthenticationService {
     this.tokenService.deleteProviderCount();
     this.limitedProfileService.deleteProfileFromSessionStorage();
     this.globalEventManagerService.setShowHeader(false);
-    let masterUiLoginUrl = this.tokenService.getMasterUiLoginUrl();
-    this.tokenService.deleteMasterUiLoginUrl();
-    this.utilityService.redirectInSameTab(masterUiLoginUrl);
+    this.goToLogin();
+  }
 
+  private goToLogin(){
+    let masterUiLoginUrl = this.tokenService.getMasterUiLoginUrl();
+    if(masterUiLoginUrl){
+      this.tokenService.deleteMasterUiLoginUrl();
+      this.utilityService.redirectInSameTab(masterUiLoginUrl);
+    }else{
+      this.router.navigate([this.LOGIN]);
+    }
   }
 
   isLogin(){
